[http2] draft typed benchmark
