--- http2/test_iri2.py
@@ -1,1 +1,0 @@
-__author__ = 'zeinamigeed'
