+++ http2/typed/README.md
@@ -1,0 +1,4 @@
+bg:
+- added class Iri2Uri, so I could use @fields to type `escape_range`
+- moved unittest cases to separate file
+- removed unittest, wrote custom `assertEqual` and `assertNotEqual`
